+++ arclight-common/src/main/java/io/izzel/arclight/common/mixin/core/world/entity/monster/ShulkerMixin.java
@@ -51,14 +51,14 @@ public abstract class ShulkerMixin extends PathfinderMobMixin {
                     if (direction != null) {
                         var event = ForgeEventFactory.onEnderTeleport((Shulker) (Object) this, blockpos1.getX(), blockpos1.getY(), blockpos1.getZ());
                         if (event.isCanceled()) direction = null;
-                        blockpos1 = new BlockPos(event.getTargetX(), event.getTargetY(), event.getTargetZ());
+                        blockpos1 = new BlockPos((int) event.getTargetX(), (int) event.getTargetY(), (int) event.getTargetZ());
                     }
                     if (direction != null) {
                         EntityTeleportEvent teleport = new EntityTeleportEvent(this.getBukkitEntity(), this.getBukkitEntity().getLocation(), new Location(((WorldBridge) this.level).bridge$getWorld(), blockpos1.getX(), blockpos1.getY(), blockpos1.getZ()));
                         Bukkit.getPluginManager().callEvent(teleport);
                         if (!teleport.isCancelled()) {
                             Location to = teleport.getTo();
-                            blockpos1 = new BlockPos(to.getX(), to.getY(), to.getZ());
+                            blockpos1 = new BlockPos((int) to.getX(), (int) to.getY(), (int) to.getZ());
                         } else {
                             return false;
                         }
